Fix laggy panning on first image if you do no zooming
The panning function tries to move the initial image in the pyramid (highest quality), whilst we could be moving the second or third (lower quality / no visible degradation!, smoother motion).

We must update the picture shown initially, to one lower in the pyramid after the pyramid has been created. This will now happen when the user tries to pan or zoom the image for the first time.

We still render the first entry in the pyramid first. I dont think waiting for the lower pyramid entries to generate, then rendering those would provide much faster initial renders, but who knows.